pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
pythonpath = ["src"]

[tool.setuptools.dynamic]
version = {attr = "asciistuff.__info__.__version__"}

[tool.setuptools.packages.find]
where = ["src"]

[project]
name = "asciistuff"
authors = [
  {name="Alexandre D'Hondt", email="alexandre.dhondt@gmail.com"},
]
description = "Library for producing ASCII arts from a text or an image"
license = {file = "LICENSE"}
keywords = ["python", "development", "programming", "ascii-art", "banner-generator", "quote-generator", "cowsay"]
requires-python = ">=3.8,<4"
classifiers = [
  "Development Status :: 5 - Production/Stable",
  "Environment :: Console",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
  "colorama",
  "cowpy>=1.1.2",
  "pillow>=10.2.0",
  "PyFiglet",
]
dynamic = ["version"]

[project.readme]
file = "README.md"
content-type = "text/markdown"

[project.urls]
documentation = "https://python-asciistuff.readthedocs.io/en/latest/?badge=latest"
homepage = "https://github.com/dhondta/python-asciistuff"
issues = "https://github.com/dhondta/python-asciistuff/issues"
repository = "https://github.com/dhondta/python-asciistuff"

[project.scripts]
asciistuff = "asciistuff:main"
